Split ffconfig into multiple files and use Cobra
